+++ projects/pioneer-charts-docs/src/app/docs/data-contract/data-contract.component.ts
@@ -59,7 +59,7 @@ export class DataContractComponent {
   barChartConfig = `import { IPcacChartConfig } from '../core/chart.model';
 
 export interface IPcacBarVerticalChartConfig extends IPcacAxisChartConfig {
-  domainMax: number;
+  isStacked: boolean;
   ...
 }`;